requirements.txt
aiogram>=3.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
colorlog>=6.7.0
aiogram-calendar==0.5.0